feat(consul): Adds Consul backend integration
Provide the APIs for configuring Vault to use consul as a secret
backend.

To make testing Vault with Consul possible, an example setup is
available within the authstore/ directory along with a README to
explain what is there and how to use it.

To make working with an existing Vault instance possible, an internal
module has been added that handles backing up and recovering the state
of the Vault instance and a public API `prepare` for reinitializing
the state from the file and information available within Vault itself.

The init() method will now return an instance of Vault if the Vault
has already been ititialized instead of throwing an error to enable
a simpler workflow when restarting an instance.

Minor bugfix to `getSealStatus` and `seal` such that each returns
a promise and supports standard promise related chain functions.